Rename status -> status_controller
