+++ server/athenian/api/__init__.py
@@ -14,7 +14,7 @@ from sentry_sdk.integrations.aiohttp import AioHttpIntegration
 from sentry_sdk.integrations.sqlalchemy import SqlalchemyIntegration
 
 from athenian.api.auth import Auth0, auth_is_configured
-from athenian.api.controllers.status import setup_status
+from athenian.api.controllers.status_controller import setup_status
 from athenian.api.metadata import __description__, __package__, __version__
 from athenian.api.serialization import FriendlyJson
 from athenian.api.slogging import add_logging_args
